+++ src/settings/derived_settings.hpp
@@ -11,12 +11,12 @@
 #pragma once
 
 namespace slide::settings {
-constexpr size_t CELL_NDATA_MAX = DATASTORE_CELL <= cellDataStorageLevel::storeHistogramData ? CELL_NDATA_HIST_MAX : CELL_NDATA_INST_MAX;
+constexpr size_t CELL_NDATA_MAX = DATASTORE_CELL <= CellDataStorageLevel::storeHistogramData ? CELL_NDATA_HIST_MAX : CELL_NDATA_INST_MAX;
 //!< length of arrays in which we store cycling data at every time step
 } // namespace slide::settings
 
 namespace slide::settings::data {
-constexpr bool storeCumulativeData = (DATASTORE_CELL >= cellDataStorageLevel::storeCumulativeData);
+constexpr bool storeCumulativeData = (DATASTORE_CELL >= CellDataStorageLevel::storeCumulativeData);
 // constexpr size_t N_CumulativeModule = (DATASTORE_MODULE >= moduleDataStorageLevel::storeCumulativeData) ? 3 : 0;
 
 } // namespace slide::settings::data